+++ src/Mscc.GenerativeAI/FilesModel.cs
@@ -22,7 +22,7 @@ namespace Mscc.GenerativeAI
         /// Lists the metadata for Files owned by the requesting project.
         /// </summary>
         /// <param name="pageSize">The maximum number of Models to return (per page).</param>
-        /// <param name="pageToken">A page token, received from a previous files.list call. Provide the pageToken returned by one request as an argument to the next request to retrieve the next page.</param>
+        /// <param name="pageToken">A page token, received from a previous ListFiles call. Provide the pageToken returned by one request as an argument to the next request to retrieve the next page.</param>
         /// <returns>List of files in File API.</returns>
         /// <exception cref="NotSupportedException">Thrown when the functionality is not supported by the model.</exception>
         /// <exception cref="HttpRequestException">Thrown when the request fails to execute.</exception>
@@ -44,22 +44,17 @@ namespace Mscc.GenerativeAI
             return await Deserialize<ListFilesResponse>(response);
         }
 
-        private void GuardSupported()
-        {
-            throw new NotImplementedException();
-        }
-
         /// <summary>
         /// Gets the metadata for the given File.
         /// </summary>
-        /// <param name="file">Required. The resource name of the file to get. This name should match a file name returned by the files.list method. Format: files/file-id.</param>
+        /// <param name="file">Required. The resource name of the file to get. This name should match a file name returned by the ListFiles method. Format: files/file-id.</param>
         /// <returns>Metadata for the given file.</returns>
-        /// <exception cref="ArgumentNullException">Thrown when the <paramref name="file"/> is null or empty.</exception>
+        /// <exception cref="ArgumentException">Thrown when the <paramref name="file"/> is <see langword="null"/> or empty.</exception>
         /// <exception cref="NotSupportedException">Thrown when the functionality is not supported by the model.</exception>
         /// <exception cref="HttpRequestException">Thrown when the request fails to execute.</exception>
         public async Task<FileResource> GetFile(string file)
         {
-            if (string.IsNullOrEmpty(file)) throw new ArgumentNullException(nameof(file));
+            if (string.IsNullOrEmpty(file)) throw new ArgumentException("Value cannot be null or empty.", nameof(file));
             // this.GuardSupported();
 
             file = file.SanitizeFileName();
@@ -74,14 +69,14 @@ namespace Mscc.GenerativeAI
         /// <summary>
         /// Deletes a file.
         /// </summary>
-        /// <param name="file">Required. The resource name of the file to get. This name should match a file name returned by the files.list method. Format: files/file-id.</param>
+        /// <param name="file">Required. The resource name of the file to get. This name should match a file name returned by the ListFiles method. Format: files/file-id.</param>
         /// <returns>If successful, the response body is empty.</returns>
-        /// <exception cref="ArgumentNullException">Thrown when the <paramref name="file"/> is null or empty.</exception>
+        /// <exception cref="ArgumentException">Thrown when the <paramref name="file"/> is <see langword="null"/> or empty.</exception>
         /// <exception cref="NotSupportedException">Thrown when the functionality is not supported by the model.</exception>
         /// <exception cref="HttpRequestException">Thrown when the request fails to execute.</exception>
         public async Task<string> DeleteFile(string file)
         {
-            if (string.IsNullOrEmpty(file)) throw new ArgumentNullException(nameof(file));
+            if (string.IsNullOrEmpty(file)) throw new ArgumentException("Value cannot be null or empty.", nameof(file));
             // this.GuardSupported();
 
             file = file.SanitizeFileName();