+++ src/Mscc.GenerativeAI/Types/ImageGeneration/ImageTextResponse.cs
@@ -8,26 +8,26 @@ namespace Mscc.GenerativeAI
         /// <summary>
         /// List of text strings representing captions, sorted by confidence.
         /// </summary>
-        public string[] Predictions { get; set; }
+        public string[]? Predictions { get; set; }
 
         /// <summary>
         /// 
         /// </summary>
-        public string DeployedModelId { get; set; }
+        public string? DeployedModelId { get; set; }
 
         /// <summary>
         /// 
         /// </summary>
-        public string Model { get; set; }
+        public string? Model { get; set; }
 
         /// <summary>
         /// 
         /// </summary>
-        public string ModelDisplayName { get; set; }
+        public string? ModelDisplayName { get; set; }
 
         /// <summary>
         /// 
         /// </summary>
-        public string ModelVersionId { get; set; }
+        public string? ModelVersionId { get; set; }
     }
 }